third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-createAnswer-expected.txt

Tests RTCPeerConnection createAnswer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testExecutionOrderClosedConnection() did not throw exception.
PASS error.name is "InvalidStateError"
PASS error.toString() is "InvalidStateError: The RTCPeerConnection's signalingState is 'closed'."
PASS events is [1,2,3]
PASS createOffer request without remote description failed.
PASS pc.setRemoteDescription(sessionDescription, descriptionSet, unexpectedCallback); did not throw exception.
PASS setRemoteDescription request succeeded.
PASS pc.createAnswer(answerCreated, unexpectedCallback); did not throw exception.
PASS createAnswer request succeeded.
PASS successfullyParsed is true

TEST COMPLETE
